requirements.txt
anyio==4.12.0
certifi==2025.11.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
llvmlite==0.44.0
numba==0.61.2
numpy==2.2.6
pandas==2.3.3
pandas-ta==0.4.71b0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
websocket-client==1.9.0
websockets==10.4
